refactor(modules): fold the MCP client core into modules/protocols/mcp/
Move the MCP transport/client layer — mcp_client.c, mcp_client_registry.c, and
their headers — from server/ and headers/ into the existing
src/modules/protocols/mcp/ module, so the MCP protocol core lives together.

This is the cycle-safe half of the MCP consolidation. These four files have no
server.h / agent / server-internal dependencies, and the module already reached
outward to mcp_client_registry.h (from mcp_tools.c), so folding the client layer
in converts that cross-directory edge into an intra-module one — strictly
improving layering.

The server-side MCP request handlers (server/server_mcp*.c and
headers/server_mcp_internal.h) are deliberately LEFT in server/: they include
server.h and are built on server_ctx_t/server_conn_t, and server.c includes
server_mcp_internal.h in return, so folding them would create a
modules/protocols/mcp <-> server cycle. They are consumers of the MCP core, not
the core itself.

Build: no C_FLAGS change (-Imodules/protocols/mcp is already present); re-prefix
mcp_client.c / mcp_client_registry.c in SERVER_SRCS (src/Makefile), the CMake
DATA_SRCS entries, and the object paths in src/tests/Rules.mk. Two path-qualified
"headers/mcp_client_registry.h" includes are made bare (agent_policy.c,
agent_tools.c).
