+++ src/modules/protocols/mcp/mcp_client_registry.c
